fix: delete company

+++ src/app/services/company-api.service.ts
@@ -20,7 +20,7 @@ export class CompanyApiService {
 
   getCompanyCreateForm(): Observable<any> {
     const API_URL = `${this.baseURL + environment.company}`;
-    return this.httpClient.get(API_URL);
+    return this.httpClient.get(API_URL)
   }
 
   createCompany(data: any): Observable<any> {
@@ -32,11 +32,17 @@ export class CompanyApiService {
     return this.httpClient.post(API_URL, data);
   }
 
-  deleteCompanyList(companyIds): Observable<any> {
-    const data = {ids: companyIds}
-    const id = companyIds[0]
-    const API_URL = `${this.baseURL + environment.company}/${id}`;
-    return this.httpClient.delete(API_URL);
+  deleteCompany(companyIds: any[]): Observable<any> {
+    if (companyIds.length > 1) {
+      const data = {ids: companyIds}
+      const API_URL = `${this.baseURL + environment.company_delete_multiple}`;
+      return this.httpClient.post(API_URL, data);
+    }
+    else {
+      const companyId = companyIds[0]
+      const API_URL = `${this.baseURL + environment.company_delete}/${companyId}`;
+      return this.httpClient.delete(API_URL);
+    }
   }
   
   subject: ReplaySubject<any> = new ReplaySubject();
